Added CRUD operations for Bookings. Created an All Bookings and My Bookings page.

+++ TravelHub.Core/Services/BookingService.cs
@@ -4,6 +4,8 @@
     using TravelHub.Core.Repositories;
     using TravelHub.Domain.Models;
     using Microsoft.EntityFrameworkCore;
+    using System.Collections.Generic;
+    using TravelHub.ViewModels.Bookings;
 
     public class BookingService : IBookingService
     {
@@ -23,17 +25,66 @@
                 BookDate = DateTime.UtcNow
             };
 
+            bool doesExist = await this.repository.All<Booking>()
+                .AnyAsync(b => b.TravelId == travelId
+                    && b.UserId == userId);
+
+            if (doesExist)
+            {
+                return;
+            }
+
             await this.repository.AddAsync(booking);
             await this.repository.SaveChangesAsync();
         }
 
         public async Task RemoveBookingAsync(int travelId, string userId)
         {
-            Booking booking = await this.repository.All<Booking>()
-                .FirstAsync(b => b.UserId == userId && b.TravelId == travelId);
+            var booking = await this.repository.All<Booking>()
+                .FirstOrDefaultAsync(b => b.UserId == userId && b.TravelId == travelId);
+
+            if (booking == null)
+            {
+                return;
+            }
 
             this.repository.Delete(booking);
             await this.repository.SaveChangesAsync();
+        }
+
+        public async Task<ICollection<BookingViewModel>> GetAllAsync()
+        {
+            return await this.repository.All<Booking>()
+                .Select(b => new BookingViewModel()
+                {
+                    Destination = $"{b.Travel.Destination.Place}, {b.Travel.Destination.Country}",
+                    ImageUrl = b.Travel.Destination.ImageUrl,
+                    Price = b.Travel.Price,
+                    Owner = $"{b.User.FirstName} {b.User.LastName} ({b.User.UserName})",
+                    DateFrom = b.Travel.DateFrom,
+                    DateTo = b.Travel.DateTo,
+                    BookDate = b.BookDate,
+                    TravelId = b.TravelId,
+                    UserId = b.UserId
+                }).ToListAsync();
+        }
+
+        public async Task<ICollection<BookingViewModel>> GetForUserAsync(string userId)
+        {
+            return await this.repository.All<Booking>()
+                .Where(b => b.UserId == userId)
+                .Select(b => new BookingViewModel()
+                {
+                    Destination = $"{b.Travel.Destination.Place}, {b.Travel.Destination.Country}",
+                    ImageUrl = b.Travel.Destination.ImageUrl,
+                    Price = b.Travel.Price,
+                    Owner = $"{b.User.FirstName} {b.User.LastName} ({b.User.UserName})",
+                    DateFrom = b.Travel.DateFrom,
+                    DateTo = b.Travel.DateTo,
+                    BookDate = b.BookDate,
+                    TravelId = b.TravelId,
+                    UserId = b.UserId
+                }).ToListAsync();
         }
     }
 }
